Solved Chapter-3 and Chapter-4 Problems

+++ Chapter-2/main.py
@@ -1,9 +1,12 @@
+# Write a python program to create files in the current directory.
+
+
 import os
 
-folder = os.path.dirname(os.path.abspath(__file__))
+folder= os.path.dirname(os.path.abspath(__file__))
 
-for i in range(1, 6):
-    filename = os.path.join(folder, f"Problem-{i}.py")
-
-    with open(filename, "w") as file:
-        pass
+for i in range(1,7):
+    filename=os.path.join(folder, f"Problem-{i}.py")
+    if not os.path.exists(filename):
+        with open(filename, "w") as file:
+            pass
